Major Update
Changelog:

Added updated loadouts.

Changed loadout addon options for ease of use.

Replaced script vehicle respawn with module due to small issues.

Added reserve maps to github, these are maps that we have used in the past or look quite nice which I have set up for jsdf to use, so if map mod is added to server they are ready to go.

Changed contents of supply crates on supply crate spawner.